Extract classic operations list from dashboard and reuse it in import
