+++ src/styles/globals.scss
@@ -1577,95 +1577,3 @@
     }
   }
 }
-
-// Operations
-// =============================================================================
-.Operations {
-  color: var(--sds-clr-gray-11);
-
-  &__note {
-    color: var(--sds-clr-gray-09);
-  }
-
-  &__filters {
-    .Button {
-      border-radius: pxToRem(16px);
-    }
-  }
-
-  // Operation type badge, e.g. "create_account".
-  &__type {
-    align-self: flex-start;
-    padding: pxToRem(2px) pxToRem(8px);
-    border: 1px solid var(--sds-clr-lilac-06);
-    border-radius: pxToRem(4px);
-    color: var(--sds-clr-lilac-11);
-    font-family: var(--sds-ff-monospace), monospace;
-  }
-
-  // Short type hint after a value, e.g. "account id".
-  &__annotation {
-    margin-left: pxToRem(6px);
-    font-size: pxToRem(12px);
-    font-weight: var(--sds-fw-regular);
-    color: var(--sds-clr-gray-09);
-  }
-
-  &__block {
-    margin: 0;
-    white-space: pre-wrap;
-    word-break: break-all;
-    font-family: var(--sds-ff-monospace), monospace;
-    font-size: pxToRem(12px);
-    line-height: pxToRem(18px);
-  }
-
-  &__cell {
-    word-break: break-word;
-
-    code {
-      font-family: var(--sds-ff-monospace), monospace;
-    }
-
-    &[data-is-label="true"] {
-      color: var(--sds-clr-gray-11);
-    }
-  }
-
-  &__gridTableContainer {
-    width: 100%;
-    overflow-x: auto;
-
-    border: 1px solid var(--sds-clr-gray-06);
-    border-radius: pxToRem(8px);
-    background-color: var(--sds-clr-base-00);
-
-    table {
-      width: 100%;
-
-      tr {
-        display: grid;
-        grid-template-columns:
-          minmax(pxToRem(140px), 1fr)
-          minmax(pxToRem(220px), 2fr);
-      }
-
-      td {
-        display: flex;
-        flex-direction: column;
-        padding: pxToRem(8px) pxToRem(12px);
-        justify-content: center;
-
-        &:not(:last-child) {
-          border-right: 1px solid var(--sds-clr-gray-06);
-        }
-      }
-
-      tbody {
-        tr:not(:first-child) {
-          border-top: 1px solid var(--sds-clr-gray-06);
-        }
-      }
-    }
-  }
-}
